backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
mssql-django==1.7
openai>=1.0.0
pyodbc==5.3.0
python-dotenv==1.2.2
requests==2.32.3
sqlparse==0.5.5
